+++ molviewspec/molviewspec/nodes.py
@@ -40,6 +40,10 @@ KindT = Literal[
 ]
 
 
+AdditionalProperties = Optional[Mapping[str, Any]]
+additional_properties_name = "additional_properties"  # make potential refactors a bit easier -- check also refs below
+
+
 class Node(BaseModel):
     """
     Base impl of all state tree nodes.
@@ -47,6 +51,9 @@ class Node(BaseModel):
 
     kind: KindT = Field(description="The type of this node.")
     params: Optional[Mapping[str, Any]] = Field(description="Optional params that are needed for this node.")
+    additional_properties: AdditionalProperties = Field(
+        description="Optional free-style dict with custom, non-schema props."
+    )
     children: Optional[list[Node]] = Field(description="Optional collection of nested child nodes.")
 
 
@@ -563,10 +570,8 @@ class ApplySelectionInlineParams(BaseModel):
     Params to customize how surroundings of a Component are presented.
     """
 
-    surroundings_radius: Optional[float] = (
-        Field(
-            description="Distance threshold in Angstrom, everything below this cutoff will be included as surroundings"
-        )
+    surroundings_radius: Optional[float] = Field(
+        description="Distance threshold in Angstrom, everything below this cutoff will be included as surroundings"
     )
     show_non_covalent_interactions: Optional[bool] = Field(
         description="Show non-covalent interactions between this component and its surroundings?"